wait for thegraph to index new application
